split config file from main
